# web/src/main/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Revet Auth API
  description: OAuth 2.1/OIDC Server API
  termsOfService: https://revethq.com/terms
  contact:
    name: Revet
    url: https://revethq.com
    email: info@revethq.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:5000
  description: Auto generated value
- url: http://0.0.0.0:5000
  description: Auto generated value
...
paths:
...
  /authorizationServers:
    get:
      tags:
      - Authorization Servers
      summary: List authorization servers with optional pagination
      parameters:
      - name: limit
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      - name: offset
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      responses:
        "200":
          description: Authorization servers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationServersResponse'
    post:
      tags:
      - Authorization Servers
      summary: Create a new authorization server
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizationServerRequest'
      responses:
        "200":
          description: Authorization server created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationServerResponse'
        "400":
          description: Invalid request
  /authorizationServers/{authorizationServerId}:
    get:
      tags:
      - Authorization Servers
      summary: Get an authorization server by ID
      parameters:
      - name: authorizationServerId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Authorization server found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationServerResponse'
        "404":
          description: Authorization server not found
    put:
      tags:
      - Authorization Servers
      summary: Update an authorization server by ID
      parameters:
      - name: authorizationServerId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizationServerRequest'
      responses:
        "200":
          description: Authorization server updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizationServerResponse'
        "404":
          description: Authorization server not found
    delete:
      tags:
      - Authorization Servers
      summary: Delete an authorization server by ID
      parameters:
      - name: authorizationServerId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Authorization server deleted successfully
        "404":
          description: Authorization server not found
...
  /scopes:
    get:
      tags:
      - Scopes
      summary: List scopes with optional filtering
      parameters:
      - name: authorizationServerIds
        in: query
        schema:
          type: array
          items:
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
            type: string
          nullable: true
      - name: limit
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      - name: offset
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      responses:
        "200":
          description: Scopes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopesResponse'
    post:
      tags:
      - Scopes
      summary: Create a new scope
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScopeRequest'
      responses:
        "200":
          description: Scope created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeResponse'
        "400":
          description: Invalid request
  /scopes/{scopeId}:
    get:
      tags:
      - Scopes
      summary: Get a scope by ID
      parameters:
      - name: scopeId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Scope found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeResponse'
        "404":
          description: Scope not found
    put:
      tags:
      - Scopes
      summary: Update a scope by ID
      parameters:
      - name: scopeId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScopeRequest'
      responses:
        "200":
          description: Scope updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeResponse'
        "404":
          description: Scope not found
    delete:
      tags:
      - Scopes
      summary: Delete a scope by ID
      parameters:
      - name: scopeId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Scope deleted successfully
        "404":
          description: Scope not found
  /templates:
    get:
      tags:
      - Templates
      summary: List templates with optional filtering
      parameters:
      - name: authorizationServerIds
        in: query
        schema:
          type: array
          items:
            format: uuid
            pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
            type: string
          nullable: true
      - name: limit
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      - name: offset
        in: query
        schema:
          format: int32
          type: integer
          nullable: true
      responses:
        "200":
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesResponse'
    post:
      tags:
      - Templates
      summary: Create a new template
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateRequest'
      responses:
        "200":
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        "400":
          description: Invalid request
  /templates/{templateId}:
    get:
      tags:
      - Templates
      summary: Get a template by ID
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Template found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        "404":
          description: Template not found
    put:
      tags:
      - Templates
      summary: Update a template by ID
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateRequest'
      responses:
        "200":
          description: Template updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        "404":
          description: Template not found
    delete:
      tags:
      - Templates
      summary: Delete a template by ID
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Template deleted successfully
        "404":
          description: Template not found
...
components:
  schemas:
...
    AuthorizationServerRequest:
      type: object
      properties:
        serverUrl:
          type: string
          nullable: true
        audience:
          type: string
          nullable: true
        clientCredentialsTokenExpiration:
          format: int32
          type: integer
          nullable: true
        authorizationCodeTokenExpiration:
          format: int32
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
    AuthorizationServerResponse:
      required:
      - name
      type: object
      properties:
        id:
          type: string
          nullable: true
        serverUrl:
          type: string
          nullable: true
        audience:
          type: string
          nullable: true
        clientCredentialsTokenExpiration:
          format: int32
          type: integer
          nullable: true
        authorizationCodeTokenExpiration:
          format: int32
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
        createdOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        updatedOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/ScopeResponse'
          nullable: true
    AuthorizationServersResponse:
      required:
      - page
      type: object
      properties:
        authorizationServers:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizationServerResponse'
          nullable: true
        page:
          type: object
          allOf:
          - $ref: '#/components/schemas/Page'
          nullable: true
...
    Metadata:
      type: object
      properties:
        properties:
          type: object
          additionalProperties: {}
          nullable: true
        identifiers:
          type: array
          items:
            $ref: '#/components/schemas/MetadataIdentifiersInner'
          nullable: true
        schemaValidations:
          type: array
          items:
            $ref: '#/components/schemas/MetadataSchemaValidationsInner'
          nullable: true
    MetadataIdentifiersInner:
      type: object
      properties:
        system:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
    MetadataSchemaValidationsInner:
      type: object
      properties:
        schemaId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        isValid:
          type: boolean
          nullable: true
        validatedOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
    OffsetDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50-04:00
    Page:
      type: object
      properties:
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
...
    ScopeRequest:
      required:
      - authorizationServerId
      - name
      type: object
      properties:
        authorizationServerId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        name:
          type: string
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
    ScopeResponse:
      type: object
      properties:
        id:
          type: string
          nullable: true
        authorizationServerId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        createdOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        updatedOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        name:
          type: string
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
    ScopesResponse:
      required:
      - page
      type: object
      properties:
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/ScopeResponse'
          nullable: true
        page:
          type: object
          allOf:
          - $ref: '#/components/schemas/Page'
          nullable: true
    TemplateRequest:
      type: object
      properties:
        authorizationServerId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        templateType:
          type: string
          nullable: true
        template:
          type: string
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
    TemplateResponse:
      type: object
      properties:
        id:
          type: string
          nullable: true
        authorizationServerId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        templateType:
          type: string
          nullable: true
        template:
          type: string
          nullable: true
        createdOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        updatedOn:
          type: string
          allOf:
          - $ref: '#/components/schemas/OffsetDateTime'
          nullable: true
        metadata:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metadata'
          nullable: true
    TemplatesResponse:
      required:
      - page
      type: object
      properties:
        templates:
          type: array
          items:
            $ref: '#/components/schemas/TemplateResponse'
          nullable: true
        page:
          type: object
          allOf:
          - $ref: '#/components/schemas/Page'
          nullable: true
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...